2021-2022/JustDanceIt/requirements.txt
numpy
python-vlc
mutagen